account/requirements.txt
psycopg2-binary==2.9.1
protobuf==3.18.1
flask==2.0.2
gunicorn==20.1.0
aiokafka==0.7.2
uvloop==0.16.0
aiopg==1.3.1
SQLAlchemy==1.4.25